rustGUI
To run the Rust program with the GTK GUI that I provided earlier, you will need to install the required dependencies and then compile and run the Rust code.

Here are the steps to follow:

Install GTK on your system. The specific installation instructions will depend on your operating system. You can find the installation instructions for various platforms in the GTK installation guide: https://www.gtk.org/docs/installations/

Create a new Rust project using Cargo. Open a terminal and run the following command to create a new Rust project

" cargo new my_gui_project  "

This will create a new Rust project in a directory named my_gui_project.

Edit the Cargo.toml file to add the gtk dependency. Open the Cargo.toml file in a text editor and add the following line under the [dependencies] section

"gtk = "0.9"

Replace the main.rs file in your project with the code I provided earlier. Copy the code and replace the contents of the main.rs file in your project directory with it.

Compile and run the Rust code. Open a terminal, navigate to the project directory (my_gui_project) and run the following command

" cargo run"

cargo run
This will compile and run the Rust code. The GTK window with the button should appear. Click the button to change the window title.

That's it! You should now be able to run the Rust program with the GTK GUI.